--- __pycache__/windows/preferencetabs/advanced_preferences.json
@@ -1,1 +1,0 @@
-c
