+++ __pycache__/windows/profileutils/narrator.py
